定数ファイルを作成してas constで定義しているものはそちらに移動
